refactor(java/parser-jvm-utils): use interface instead of a class to fit maven reflection
